modularize s3-unspool CLI
