+++ crates/s3-unspool-cli/src/commands/mod.rs
@@ -1,0 +1,5 @@
+mod unzip;
+mod zip;
+
+pub(crate) use unzip::run_unzip;
+pub(crate) use zip::run_zip;
