+++ crates/s3-unspool-cli/src/reports/unzip/mod.rs
@@ -1,0 +1,6 @@
+mod details;
+mod report;
+mod transcript;
+
+pub(crate) use report::UnzipCommandReport;
+pub(crate) use transcript::unzip_transcript;
